cleaned up student settings and added in tabs
