+++ client/src/components/Settings/StudentSettings/DetailsSettings.jsx
@@ -3,26 +3,21 @@ import PropTypes from 'prop-types';
 import styled from 'styled-components';
 import { toast } from 'react-toastify';
 import { useStateValue } from 'react-conflux';
-import { MaskedInput, Select, Box, Heading } from 'grommet';
-import { Trash } from 'grommet-icons';
+import { MaskedInput, Select, Box, Heading, Tab, Tabs } from 'grommet';
 
-import queries from './queries';
+import queries from '../queries';
 import {
   BaseForm,
   BaseTextInput,
   BaseFormField,
   BaseButton
-} from '../../styles/themes';
-import { globalContext } from '../../store/reducers/globalReducer';
-import c from '../../store/constants';
-import {
-  studentContext,
-  STUDENT_EMAIL_UPDATE,
-  REMOVE_STUDENT_EMAIL
-} from '../../store/reducers/studentReducer';
-import ConfirmationLayer from '../../components/ConfirmationLayer';
+} from '../../../styles/themes';
+import { globalContext } from '../../../store/reducers/globalReducer';
+import c from '../../../store/constants';
+import { studentContext } from '../../../store/reducers/studentReducer';
+import EmailSettings from './EmailSettings';
 
-const StudentSettings = ({ history }) => {
+const DetailSettings = ({ history }) => {
   const [{ user }] = useStateValue(globalContext);
   const [studentState, studentDispatch] = useStateValue(studentContext);
   const [input, setInput] = useState({
@@ -37,41 +32,12 @@ const StudentSettings = ({ history }) => {
     phone: studentState.studentData.studentDetails.phone,
     userId: user.id
   });
-  const [email, setEmail] = useState('');
-
-  const [
-    hasActiveConfirmationDialog,
-    setHasActiveConfirmationDialog
-  ] = useState(false);
 
   const handleChanges = e => {
     setInput({
       ...input,
       [e.target.name]: e.target.value
     });
-  };
-
-  const submitEmail = async e => {
-    e.preventDefault();
-    try {
-      const { data } = await queries.addUserEmail({
-        userId: user.id,
-        email
-      });
-      studentDispatch({
-        type: STUDENT_EMAIL_UPDATE,
-        payload: data.addUserEmail
-      });
-      toast.success(`Email added succesfully`, {
-        className: 'status-ok',
-        position: toast.POSITION.BOTTOM_CENTER,
-        hideProgressBar: true,
-        autoClose: true
-      });
-      setEmail('');
-    } catch (err) {
-      console.error(err);
-    }
   };
 
   const handleSubmit = async e => {
@@ -93,66 +59,8 @@ const StudentSettings = ({ history }) => {
     }
   };
 
-  const confirmRemoveEmail = async id => {
-    try {
-      await queries.deleteUserEmail({ id });
-      toast.success(`Email is deleted`, {
-        className: 'status-ok',
-        position: toast.POSITION.BOTTOM_CENTER,
-        hideProgressBar: true,
-        autoClose: true
-      });
-      const updateEmailList = studentState.studentData.studentDetails.emailList.filter(
-        email => {
-          return email.id !== id;
-        }
-      );
-      studentDispatch({ type: REMOVE_STUDENT_EMAIL, payload: updateEmailList });
-    } catch (err) {
-      console.error(err);
-    }
-  };
-
   return (
     <>
-      <StudentForm onSubmit={submitEmail}>
-        <Box direction="column">
-          <Heading margin="20px 0 0 0" alignSelf="center">
-            Add an Email
-          </Heading>
-          <StudentFormField label="Email">
-            <StudentBaseTextInput
-              name="email"
-              placeholder="fakeemail@email.com"
-              onChange={e => setEmail(e.target.value)}
-              value={email}
-              plain={false}
-            />
-          </StudentFormField>
-          <StudentButton
-            type="submit"
-            primary
-            label="Add Email"
-            alignSelf="center"
-          />
-        </Box>
-      </StudentForm>
-      <EmailBox direction="column">
-        <h2>Email List</h2>
-        {studentState.studentData.studentDetails.emailList.map(emailObj => (
-          <EmailContainer>
-            <p>{emailObj.email}</p>
-            {hasActiveConfirmationDialog && (
-              // yesFunc for when the "Yes" button is clicked; noFunc for when the "No" button is clicked (both are optional)
-              <ConfirmationLayer
-                onClose={() => setHasActiveConfirmationDialog(false)} // Needed to make the layer disappear
-                yesFunc={() => confirmRemoveEmail(emailObj.id)}
-              />
-            )}
-            <TrashButton onClick={() => setHasActiveConfirmationDialog(true)} />{' '}
-          </EmailContainer>
-        ))}
-      </EmailBox>
       <StudentForm onSubmit={handleSubmit}>
         <Box direction="column">
           <Heading margin="20px 0 0 0" alignSelf="center">
@@ -266,7 +174,7 @@ const StudentSettings = ({ history }) => {
   );
 };
 
-StudentSettings.propTypes = {
+DetailSettings.propTypes = {
   history: PropTypes.object.isRequired // eslint-disable-line react/forbid-prop-types
 };
 
@@ -300,35 +208,4 @@ const StudentMaskedInput = styled(MaskedInput)`
   /* border: ${({ theme }) => theme.global.border}; */
 `;
 
-const TrashButton = styled(Trash)`
-  cursor: pointer;
-`;
-
-const EmailBox = styled(Box)`
-  margin: 12px auto 0px;
-  max-width: 800px;
-`;
-
-const EmailContainer = styled.section`
-  margin: 10px auto 0;
-  max-width: 800px;
-  width: 100%;
-  background: white;
-  /* min-height: calc(100vh - 170px); */
-  -webkit-box-shadow: -2px 5px 25px -17px rgba(0, 0, 0, 0.61);
-  -moz-box-shadow: -2px 5px 25px -17px rgba(0, 0, 0, 0.61);
-  box-shadow: -2px 5px 25px -17px rgba(0, 0, 0, 0.61);
-  border-radius: 2px;
-  display: flex;
-  justify-content: space-between;
-  padding: 20px 3%;
-  transition: box-shadow 0.5s;
-
-  :hover {
-    -webkit-box-shadow: 0px 0px 15px -2px rgba(125, 76, 219, 1);
-    -moz-box-shadow: 0px 0px 15px -2px rgba(125, 76, 219, 1);
-    box-shadow: 0px 0px 15px -2px rgba(125, 76, 219, 1);
-  }
-`;
-
-export default StudentSettings;
+export default DetailSettings;
